NG-57 <매물 등록시 좌표 정보, 아이디 정보 추가>
