+++ src/main/java/com/kdt/controllers/EstateController.java
@@ -5,7 +5,9 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -14,10 +16,10 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.SessionAttributes;
 import org.springframework.web.multipart.MultipartFile;
 
-import com.kdt.dto.EstateDTO;
+import com.kdt.dto.UploadEstateDTO;
 import com.kdt.dto.EstateOptionDTO;
 import com.kdt.dto.EstateRequestDTO;
-import com.kdt.dto.EstateTempDTO;
+import com.kdt.dto.EstateDTO;
 import com.kdt.services.EstateService;
 
 import jakarta.servlet.http.HttpSession;
@@ -35,20 +37,24 @@ public class EstateController {
 
 	@PostMapping
 	@RequestMapping("estateInsert1")
-	public ResponseEntity<Void> insert1(@RequestBody EstateDTO dto) throws Exception {
+	public ResponseEntity<Void> insert1(@RequestBody UploadEstateDTO dto) throws Exception {
+		dto.setWriter("test1234");
 		session.setAttribute("estateDTO", dto);
-
+		
+		System.out.println(dto.getLatitude());
+		System.out.println(dto.getLongitude());
+		
 		return ResponseEntity.ok().build();
 	}
 
 	@PostMapping
 	@RequestMapping("estateInsert2")
 	public ResponseEntity<Void> insert2(@RequestBody EstateRequestDTO requestDTO) throws Exception {
-		EstateDTO dto = requestDTO.getEstateDTO();
+		UploadEstateDTO dto = requestDTO.getEstateDTO();
 		String[] optionList = requestDTO.getOptionList();
 
 		// 매물 ->
-		EstateDTO estateDTO = (EstateDTO) session.getAttribute("estateDTO");
+		UploadEstateDTO estateDTO = (UploadEstateDTO) session.getAttribute("estateDTO");
 
 		estateDTO.setTransactionCode(dto.getTransactionCode());
 		estateDTO.setDeposit(dto.getDeposit());
@@ -77,7 +83,7 @@ public class EstateController {
 			@RequestParam("title") String title, @RequestParam("contents") String contents,
 			@RequestParam("memo") String memo) throws Exception {
 
-		EstateDTO estateDTO = (EstateDTO) session.getAttribute("estateDTO");
+		UploadEstateDTO estateDTO = (UploadEstateDTO) session.getAttribute("estateDTO");
 		List<EstateOptionDTO> optionDTOList = (List<EstateOptionDTO>) session.getAttribute("optionDTOList");
 
 		estateDTO.setTitle(title);
@@ -90,10 +96,17 @@ public class EstateController {
 	}
 
 	@GetMapping
-	public ResponseEntity<List<EstateTempDTO>> selectAll() {
-		List<EstateTempDTO> list = eServ.selectAll();
+	public ResponseEntity<List<EstateDTO>> selectAll() {
+		List<EstateDTO> list = eServ.selectAll();
 		
 		return ResponseEntity.ok(list);
+	}
+	
+	@DeleteMapping("/{estateId}")
+	public ResponseEntity<Void> delete(@PathVariable Long estateId) {
+		eServ.deleteById(estateId);
+		
+		return ResponseEntity.ok().build();
 	}
 
 }
