+++ src/main/java/com/kdt/domain/entities/Estate.java
@@ -7,6 +7,8 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
 
 @Entity
@@ -17,21 +19,6 @@ public class Estate {
 	@Column(name="estate_id")
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long estateId;
-	
-	@Column(name="room_code")
-	private String roomCode;
-	
-	@Column(name="structure_code")
-	private String structureCode;
-	
-	@Column(name="building_code")
-	private String buildingCode;
-	
-	@Column(name="transaction_code")
-	private String transactionCode;
-	
-	@Column(name="heating_code")
-	private String heatingCode;
 	
 	@Column(name="deposit")
 	private Long deposit;
@@ -68,6 +55,26 @@ public class Estate {
 	
 	@Column(name="write_date")
 	private Timestamp writeDate;
+	
+	@OneToOne
+    @JoinColumn(name = "room_code", referencedColumnName = "room_id")
+    private Room room;
+	
+	@OneToOne
+    @JoinColumn(name = "structure_code", referencedColumnName = "structure_id")
+    private Structure structure;
+	
+	@OneToOne
+    @JoinColumn(name = "building_code", referencedColumnName = "building_id")
+    private Building building;
+	
+	@OneToOne
+    @JoinColumn(name = "transaction_code", referencedColumnName = "transaction_id")
+    private Transaction transaction;
+	
+	@OneToOne
+    @JoinColumn(name = "heating_code", referencedColumnName = "heating_id")
+    private HeatingSystem heatingSystem;
 
 	public Long getEstateId() {
 		return estateId;
@@ -75,46 +82,6 @@ public class Estate {
 
 	public void setEstateId(Long estateId) {
 		this.estateId = estateId;
-	}
-
-	public String getRoomCode() {
-		return roomCode;
-	}
-
-	public void setRoomCode(String roomCode) {
-		this.roomCode = roomCode;
-	}
-
-	public String getStructureCode() {
-		return structureCode;
-	}
-
-	public void setStructureCode(String structureCode) {
-		this.structureCode = structureCode;
-	}
-
-	public String getBuildingCode() {
-		return buildingCode;
-	}
-
-	public void setBuildingCode(String buildingCode) {
-		this.buildingCode = buildingCode;
-	}
-
-	public String getTransactionCode() {
-		return transactionCode;
-	}
-
-	public void setTransactionCode(String transactionCode) {
-		this.transactionCode = transactionCode;
-	}
-
-	public String getHeatingCode() {
-		return heatingCode;
-	}
-
-	public void setHeatingCode(String heatingCode) {
-		this.heatingCode = heatingCode;
 	}
 
 	public Long getDeposit() {
@@ -213,17 +180,52 @@ public class Estate {
 		this.writeDate = writeDate;
 	}
 
-	public Estate(Long estateId, String roomCode, String structureCode, String buildingCode, String transactionCode,
-			String heatingCode, Long deposit, Long price, float area, Long zipcode, String address, Long roomFloors,
-			Long buildingFloors, Long maintenanceCost, String title, String contents, String memo,
-			Timestamp writeDate) {
+	public Room getRoom() {
+		return room;
+	}
+
+	public void setRoom(Room room) {
+		this.room = room;
+	}
+
+	public Structure getStructure() {
+		return structure;
+	}
+
+	public void setStructure(Structure structure) {
+		this.structure = structure;
+	}
+
+	public Building getBuilding() {
+		return building;
+	}
+
+	public void setBuilding(Building building) {
+		this.building = building;
+	}
+
+	public Transaction getTransaction() {
+		return transaction;
+	}
+
+	public void setTransaction(Transaction transaction) {
+		this.transaction = transaction;
+	}
+
+	public HeatingSystem getHeatingSystem() {
+		return heatingSystem;
+	}
+
+	public void setHeatingSystem(HeatingSystem heatingSystem) {
+		this.heatingSystem = heatingSystem;
+	}
+
+	public Estate(Long estateId, Long deposit, Long price, float area, Long zipcode, String address,
+			Long roomFloors, Long buildingFloors, Long maintenanceCost, String title, String contents, String memo,
+			Timestamp writeDate, Room room, Structure structure, Building building, Transaction transaction,
+			HeatingSystem heatingSystem) {
 		super();
 		this.estateId = estateId;
-		this.roomCode = roomCode;
-		this.structureCode = structureCode;
-		this.buildingCode = buildingCode;
-		this.transactionCode = transactionCode;
-		this.heatingCode = heatingCode;
 		this.deposit = deposit;
 		this.price = price;
 		this.area = area;
@@ -236,11 +238,15 @@ public class Estate {
 		this.contents = contents;
 		this.memo = memo;
 		this.writeDate = writeDate;
+		this.room = room;
+		this.structure = structure;
+		this.building = building;
+		this.transaction = transaction;
+		this.heatingSystem = heatingSystem;
 	}
 
 	public Estate() {
 		super();
 	}
-	
 	
 }
