+++ src/main/java/com/kdt/dto/EstateRequestDTO.java
@@ -1,13 +1,13 @@
 package com.kdt.dto;
 
 public class EstateRequestDTO {
-	private EstateDTO estateDTO;
+	private UploadEstateDTO estateDTO;
 	private String[] optionList;
 	
-	public EstateDTO getEstateDTO() {
+	public UploadEstateDTO getEstateDTO() {
 		return estateDTO;
 	}
-	public void setEstateDTO(EstateDTO estateDTO) {
+	public void setEstateDTO(UploadEstateDTO estateDTO) {
 		this.estateDTO = estateDTO;
 	}
 	public String[] getOptionList() {
@@ -16,7 +16,7 @@ public class EstateRequestDTO {
 	public void setOptionList(String[] optionList) {
 		this.optionList = optionList;
 	}
-	public EstateRequestDTO(EstateDTO estateDTO, String[] optionList) {
+	public EstateRequestDTO(UploadEstateDTO estateDTO, String[] optionList) {
 		super();
 		this.estateDTO = estateDTO;
 		this.optionList = optionList;
