+++ src/main/java/com/kdt/repositories/UploadEstateRepository.java
@@ -3,7 +3,6 @@ package com.kdt.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.kdt.domain.entities.UploadEstate;
-import com.kdt.domain.entities.EstateImage;
 
-public interface EstateImageRepository extends JpaRepository<EstateImage, Long>{
+public interface UploadEstateRepository extends JpaRepository<UploadEstate, Long>{
 }
